fix: PITCH 기능 요청 시 반환하는 JSON 필드 값 정리 #11
- 응답에 필요한 값들만 반환 하도록 수정

+++ be/src/main/java/kr/codesquad/baseball/dto/playerVO/Batter.java
@@ -9,7 +9,7 @@ public class Batter extends PlayerVO {
 
     private int battingOrder;
 
-    private int battingAverage;
+    private double battingAverage;
 
     private int plateAppearance;
